refactor: extract games_commands.py from games.py
Move slash command implementation (_parse_slash_command,
_execute_command, _HELP_TEXT, _build_character_response,
_build_relationships_response, _execute_end_command,
_generate_epilogue, _dimension_label, _KNOWN_COMMANDS)
to games_commands.py.

games.py: 1551 → 1200 lines (-351, -22.6%)
Total decomposition: 1610 → 1200 (games.py) + 79 (helpers) + 374 (commands)
